Refactor core and component types in `wasmparser`.
This commit refactors `wasmparser` such that types related to parsing core
modules and components are separated.

This also puts the state logic for validation into a `core` and `component`
submodule.
