+++ crates/wasmparser/src/readers/component.rs
@@ -1,0 +1,13 @@
+mod aliases;
+mod exports;
+mod functions;
+mod imports;
+mod instances;
+mod types;
+
+pub use self::aliases::*;
+pub use self::exports::*;
+pub use self::functions::*;
+pub use self::imports::*;
+pub use self::instances::*;
+pub use self::types::*;
