+++ crates/wasmparser/src/readers/component/aliases.rs
