+++ crates/wasmparser/src/readers/component/instances.rs
