+++ crates/wasmparser/src/readers/core.rs
@@ -1,0 +1,35 @@
+mod code;
+mod data;
+mod elements;
+mod exports;
+mod functions;
+mod globals;
+mod imports;
+mod init;
+mod linking;
+mod memories;
+mod names;
+mod operators;
+mod producers;
+mod relocs;
+mod tables;
+mod tags;
+mod types;
+
+pub use self::code::*;
+pub use self::data::*;
+pub use self::elements::*;
+pub use self::exports::*;
+pub use self::functions::*;
+pub use self::globals::*;
+pub use self::imports::*;
+pub use self::init::*;
+pub use self::linking::*;
+pub use self::memories::*;
+pub use self::names::*;
+pub use self::operators::*;
+pub use self::producers::*;
+pub use self::relocs::*;
+pub use self::tables::*;
+pub use self::tags::*;
+pub use self::types::*;
