+++ crates/wasmparser/src/readers/core/code.rs
