+++ crates/wasmparser/src/readers/core/data.rs
