+++ crates/wasmparser/src/readers/core/globals.rs
