+++ crates/wasmparser/src/readers/core/init.rs
