+++ crates/wasmparser/src/readers/core/linking.rs
