+++ crates/wasmparser/src/readers/core/memories.rs
