+++ crates/wasmparser/src/readers/core/operators.rs
