+++ crates/wasmparser/src/readers/core/producers.rs
