+++ crates/wasmparser/src/readers/core/tables.rs
