+++ crates/wasmparser/src/readers/core/tags.rs
